+++ __tests__/config/config.notReady.js
@@ -7,8 +7,6 @@
 
 	const appName = 'Test Config';
 
-	let originalTimeout;
-
 	describe('generator-cozen-angular:config', () => {
 		beforeAll(() => {
 			return helpers.run(path.join(__dirname, '../../generators/config'))
@@ -18,15 +16,6 @@
 				.on('error', function ($error) {
 					console.log('Error !', $error);
 				});
-		});
-
-		beforeEach(() => {
-			originalTimeout                  = jasmine.DEFAULT_TIMEOUT_INTERVAL;
-			jasmine.DEFAULT_TIMEOUT_INTERVAL = 30000;
-		});
-
-		afterEach(() => {
-			jasmine.DEFAULT_TIMEOUT_INTERVAL = originalTimeout;
 		});
 
 		it('is not ready');