+++ __tests__/filter/filter.notReady.js
@@ -5,8 +5,6 @@
 	const path    = require('path');
 	const helpers = require('yeoman-test');
 
-	let originalTimeout;
-
 	describe('generator-cozen-angular:filter', () => {
 		beforeAll(() => {
 			return helpers.run(path.join(__dirname, '../../generators/filter'))
@@ -16,15 +14,6 @@
 				.on('error', function ($error) {
 					console.log('Error !', $error);
 				});
-		});
-
-		beforeEach(() => {
-			originalTimeout                  = jasmine.DEFAULT_TIMEOUT_INTERVAL;
-			jasmine.DEFAULT_TIMEOUT_INTERVAL = 30000;
-		});
-
-		afterEach(() => {
-			jasmine.DEFAULT_TIMEOUT_INTERVAL = originalTimeout;
 		});
 
 		it('is not ready');