+++ __tests__/view/view.newUser.js
@@ -8,8 +8,6 @@
 
 	const appName = 'Test View for new user';
 
-	let originalTimeout;
-
 	describe('generator-cozen-angular:view', () => {
 		beforeAll(() => {
 			return helpers.run(path.join(__dirname, '../../generators/view'))
@@ -21,15 +19,6 @@
 					authorEmail    : 'geoffrey.testelin@gmail.com',
 					stateAbstract  : true
 				});
-		});
-
-		beforeEach(() => {
-			originalTimeout                  = jasmine.DEFAULT_TIMEOUT_INTERVAL;
-			jasmine.DEFAULT_TIMEOUT_INTERVAL = 30000;
-		});
-
-		afterEach(() => {
-			jasmine.DEFAULT_TIMEOUT_INTERVAL = originalTimeout;
 		});
 
 		it('create view js files', () => {